deps: Migrate to path-extra library for handling file paths
`path-extra' reexports `path', also add path-io. `text' is a dependency
of `path' so add it do use T.pack.

`path-extra' provides ~ and $ENV_VAR parsing on top of `path'. It
isn't on hackage so install it from git using `cabal.project'.

also make the default file Maybe and default to Nothing. Get the
default config file path using the renamed function getDefConfigFile,
defaults to <current directory>/torange-bot.conf.

Also note that the use of QuasiQuotes/Template Haskell is also a
"dependency" of the `path' library because it handles multi-platform
file path encoding, so it can't just be `String' (or `Text').